Move nginx service to end of recipe
Service resources generally should be at the end of recipes when they
have configuration files that can modify their behavior. If a
configuration file is broken, and the service fails to restart, it
will always fail to start before Chef has a chance to configure the
file/template that could fix whatever configuration was broken.